refactor DB tests
